RUBY-2149 separate resume tests from others
